pyproject.toml
[project]
name = "llm-goose"
version = "0.1.22"
description = "LLM testing framework for validating agent behavior and tool usage"
license = "MIT"
authors = [{ name = "Rafał Łazicki", email = "rafal.lazicki99@gmail.com" }]
keywords = [
    "llm",
    "testing",
    "agent",
    "langchain",
    "openai",
    "pytest",
    "ai",
    "framework",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Environment :: Console",
    "Environment :: Web Environment",
]
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.4",
    "langchain-openai>=0.1.0",
    "openai>=2.7.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.0.1",
    "typer>=0.12.5",
]

[project.optional-dependencies]
api = ["fastapi>=0.115.0", "uvicorn>=0.32.0", "websockets>=15.0.1"]

[project.urls]
Homepage = "https://github.com/Raff-dev/goose"
Repository = "https://github.com/Raff-dev/goose"
Issues = "https://github.com/Raff-dev/goose/issues"
Documentation = "https://github.com/Raff-dev/goose#readme"

[dependency-groups]
core = [
    "langchain>=1.0.4",
    "langchain-openai>=0.1.0",
    "openai>=2.7.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.0.1",
]
example = ["django>=5.2.8", "psycopg[binary]>=3.2.12"]
dev = [
    "absolufy-imports>=0.3.1",
    "pre-commit>=4.4.0",
    "black>=25.9.0",
    "autoflake>=2.3.1",
    "isort>=7.0.0",
    "pyupgrade>=3.21.1",
    "pylint>=4.0.2",
    "mypy>=1.18.2",
    "twine>=6.2.0",
    "django-stubs[compatible-mypy]>=5.2.7",
    "pytest>=9.0.1",
    "coverage>=7.12.0",
]
all = [
    { include-group = "core" },
    { include-group = "example" },
    { include-group = "dev" },
]

[project.scripts]
goose = "goose.cli:app"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["goose*"]

[tool.setuptools.package-data]
goose = ["**/*.py"]

[tool.black]
line-length = 120
target-version = ["py313"]

[tool.ruff]
line-length = 120
target-version = "py313"


[tool.pylint.format]
max-line-length = 120

[tool.bandit]
skips = ["B101"]

[tool.isort]
profile = "black"
line_length = 120
include_trailing_comma = true
multi_line_output = 3

[tool.mypy]
python_version = "3.13"
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "example_system.settings"

[tool.autoflake]
in-place = true
remove-all-unused-imports = true
remove-unused-variables = true
ignore-init-module-imports = true

[tool.pylint.main]
max-line-length = 120
disable = ["too-few-public-methods", "missing-module-docstring", "fixme", "duplicate-code"]
ignore = ["migrations", "tests", "template"]
ignore-patterns = [".*/migrations/.*", "tests/.*", ".*/template/.*"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
filterwarnings = [
    "ignore:AgentStatePydantic has been moved to `langchain.agents`:Warning",
]

[tool.coverage.run]
branch = true
source = ["goose"]
omit = ["example_system/*", "goose/cli.py", "goose/api/cli.py", "goose/testing/cli.py"]

[tool.coverage.report]
show_missing = true
skip_covered = true
